+++ include/caffe/net.hpp
@@ -195,8 +195,8 @@ class Net {
     return param_display_names_;
   }
   /// @brief Input and output blob numbers
-  inline int num_inputs() const { return net_input_blobs_.size(); }
-  inline int num_outputs() const { return net_output_blobs_.size(); }
+  inline int num_inputs() const { return static_cast<int>(net_input_blobs_.size()); }
+  inline int num_outputs() const { return static_cast<int>(net_output_blobs_.size()); }
   inline const vector<Blob<Dtype>*>& input_blobs() const {
     return net_input_blobs_;
   }
