+++ src/caffe/layers/input_layer.cpp
@@ -7,7 +7,7 @@ namespace caffe {
 template <typename Dtype>
 void InputLayer<Dtype>::LayerSetUp(const vector<Blob<Dtype>*>& bottom,
       const vector<Blob<Dtype>*>& top) {
-  const int num_top = top.size();
+  const int num_top = static_cast<int>(top.size());
   const InputParameter& param = this->layer_param_.input_param();
   const int num_shape = param.shape_size();
   CHECK(num_shape == 0 || num_shape == 1 || num_shape == num_top)
